areta/requirements.txt
cachetools==4.1.1
chardet==3.0.4
convert-numbers==0.4
docopt==0.6.2
editdistance==0.5.3
edlib==1.3.8.post2
joblib==0.17.0
matplotlib==3.3.3
nltk==3.6.6
numpy==1.21.0
openpyxl==3.0.5
pandas==1.1.4
pandas-ml==0.6.1
prettytable==0.7.2
PyArabic==0.6.10
pytest==6.1.1
python-editor==1.0.4
python-Levenshtein==0.12.0
regex==2022.3.2
scikit-learn==0.23.2
scipy==1.5.2
tqdm==4.50.2
xlrd==2.0.1
